refactor: 상수 및 커스텀 예외처리 적용
