+++ src/main/java/gift/exception/oauth/KakaoTokenNotFoundException.java
@@ -1,0 +1,11 @@
+package gift.exception.oauth;
+
+import gift.exception.KakaoOAuthException;
+import org.springframework.http.HttpStatus;
+
+public class KakaoTokenNotFoundException extends KakaoOAuthException {
+
+    public KakaoTokenNotFoundException(String message) {
+        super(message, HttpStatus.BAD_REQUEST);
+    }
+}
