--- src/main/java/gift/util/constants/TokenConstants.java
@@ -1,6 +1,0 @@
-package gift.util.constants;
-
-public class TokenConstants {
-
-    public static final String TOKEN_NOT_FOUND = "토큰을 찾을 수 없습니다.";
-}
