+++ src/main/java/gift/util/constants/auth/KakaoOAuthConstants.java
@@ -1,4 +1,4 @@
-package gift.util.constants;
+package gift.util.constants.auth;
 
 public class KakaoOAuthConstants {
 
@@ -9,4 +9,5 @@ public class KakaoOAuthConstants {
     public static final String SCOPES_FAILURE_ERROR = "동의 내역 가져오기 요청이 실패했습니다: ";
     public static final String USERINFO_RESPONSE_ERROR = "사용자 정보 응답이 올바르지 않습니다.";
     public static final String USERINFO_FAILURE_ERROR = "사용자 정보 가져오기 요청이 실패했습니다: ";
+    public static final String TOKEN_NOT_FOUND = "카카오 토큰을 찾을 수 없습니다.";
 }
