+++ src/main/java/gift/util/constants/auth/TokenConstants.java
@@ -1,0 +1,7 @@
+package gift.util.constants.auth;
+
+public class TokenConstants {
+
+    public static final String EXPIRED_TOKEN = "토큰이 만료되었습니다.";
+    public static final String INVALID_TOKEN = "유효하지 않은 토큰입니다.";
+}
